src/requirements.txt
pyyaml
scipy
fastapi
uvicorn
requests
websocket_client
pandas
pyyaml
pyotp
docker
schedule
websockets
asyncio
websocket_server
nsepythonserver
sqlalchemy
psycopg2-binary
setuptools
pandas
redis
mibian
apscheduler
dhanhq
Dependencies/NorenRestApi-0.0.30-py2.py3-none-any.whl